Add files via upload
The Library Management System is a database-based project designed to manage books, members, and book issue-return transactions efficiently. It keeps track of all library resources such as book details, member information, and the records of books that are issued or returned.

This system helps librarians and administrators easily maintain data, reduce manual work, and ensure accurate book tracking. It uses SQL to create and manage tables (Books, Members, and Issue_Return) and to perform operations such as insertion, updating, and retrieval of data. The project ensures data consistency and relationship through primary and foreign keys.